Cambio de lugar de SearchBar a Carpeta NavBar
