+++ client/src/Components/NavBar/SearchBar/SearchBar.module.css
@@ -1,0 +1,67 @@
+/* COLORES DE PRUEBA MODIFICAR DESPUES */
+
+.containSearchBar {
+   margin-top: 40px;
+   /* quitar este margin-top si es necesario*/
+   text-align: center;
+   display: flex;
+   /* justify-items: flex-start; */
+   align-items: center;
+   justify-content: center;
+   /* flex-direction: column; */
+   margin-bottom: 70px;
+}
+
+.mainInput {
+   width: 230px;
+   margin: 0 5px;
+   /* margin-bottom: 10px; */
+   /* background-color: aqua; */
+}
+
+.mainInputContainer {
+   width: 280px;
+   border-radius: 100px;
+   border: 1px solid #dfe1e5;
+   display: flex;
+   justify-content: center;
+   align-items: center;
+   transition-delay: all 6s;
+}
+
+.mainInputContainer:hover {
+   box-shadow: 0 1px 6px 0 #20212447;
+   border-color: #dfe1e500;
+}
+.mainInput input {
+   width: 200px;
+   height: 40px;
+   border: none;
+   outline: none;
+}
+
+/*    BUTTON     */
+
+.mainButtons {
+   width: 230px;
+   margin: 0 10px;
+}
+.mainButtons button {
+   height: 36px;
+   background-color: #efb810;
+   border: 2px solid #efb810;
+   border-radius: 4px;
+   font-size: 1.4rem;
+   font-weight: bold;
+   text-align: justify;
+   font-family: "DM-Sans", sans-serif;
+   color: #1c1c1c;
+   padding: 0 15px;
+}
+
+.mainButtons button:hover {
+   background-color: rgb(223, 169, 23);
+   cursor: pointer;
+   color: #1c1c1c;
+   border: 1px solid #1c1c1c;
+}
